Add files via upload
Task-1
Use the provided Customer and Order CSV files to create corresponding tables in a database.
Task-2
Build a simple API on top of this data. This API should provide 2 REST endpoints that return JSON for Customers and their Orders.
The REST endpoints should be as follows:
● One that returns Customer and Order data for all Customers. This endpoint should also allow you to query for all active or archived Customers.
● Another that returns the same data for a single Customer using a Customer id.
Task-3
Build an Integration that could be run on a schedule, for example as an ETL job. This should call the first endpoint of your API above to:
● Return the JSON payload for all active customers.
● Then transform the data by concatenating “firstname” and “surname” fields into a single “name” field.
● Then send all the data onwards to a target API that only accepts a payload with a single customer record.
● Then log the returned HTTP Response Code.